Move RegionManagerExtensions.AddToRegion extension method to IRegionManager.
